requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
websockets>=12.0
httpx>=0.27.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
watchdog>=4.0.0
tavily-python>=0.3.0
pillow>=10.0.0
